Create extension object for jpos plugin
The plugin can now be configured using a 'jpos' directive in the .gradle file:

```groovy
jpos {
    target = "devel"
    addGitRevision = true
    addBuildTime = true
    archiveJarName = "${project_name}-${version}.jar"
    archiveWarName = "${project_name}-${version}.war"
    installDir = build/install/${project_name}
    distDir = src/dist
}
```

A non trivial refactor of the plugin was made to ensure that those properties
can be configured and used as described in the README.

This commit also makes this plugin compatible with gradle 8, so it fixes #3

The configurability of git revision data also fixes #5.

And it can be a workaround for #7 using a different distDir.

Signed-off-by: Arturo Volpe <avolpe@fintech.works>